Fix tests in PIT.
