+++ .gitignore
@@ -11,3 +11,4 @@ release.properties
 .DS_Store
 .idea
 /package-lock.json
+metrics.env
