+++ doc/dependency-graph.puml
@@ -9,13 +9,13 @@ skinparam rectangle {
 rectangle "jackson-databind\n\n2.16.0" as com_fasterxml_jackson_core_jackson_databind_jar
 rectangle "jackson-annotations\n\n2.16.0" as com_fasterxml_jackson_core_jackson_annotations_jar
 rectangle "jackson-core\n\n2.16.0" as com_fasterxml_jackson_core_jackson_core_jar
-rectangle "autograding-model\n\n3.9.0" as edu_hm_hafner_autograding_model_jar
+rectangle "autograding-model\n\n3.11.0" as edu_hm_hafner_autograding_model_jar
 rectangle "autograding-github-action\n\n3.8.0-SNAPSHOT" as edu_hm_hafner_autograding_github_action_jar
 rectangle "analysis-model\n\n11.13.0" as edu_hm_hafner_analysis_model_jar
 rectangle "jsoup\n\n1.16.2" as org_jsoup_jsoup_jar
 rectangle "commons-digester3\n\n3.2" as org_apache_commons_commons_digester3_jar
 rectangle "cglib\n\n2.2.2" as cglib_cglib_jar
-rectangle "commons-logging\n\n1.2" as commons_logging_commons_logging_jar
+rectangle "commons-logging\n\n1.3.0" as commons_logging_commons_logging_jar
 rectangle "commons-beanutils\n\n1.9.4" as commons_beanutils_commons_beanutils_jar
 rectangle "commons-collections\n\n3.2.2" as commons_collections_commons_collections_jar
 rectangle "commons-text\n\n1.10.0" as org_apache_commons_commons_text_jar
@@ -48,7 +48,7 @@ rectangle "commons-lang3\n\n3.14.0" as org_apache_commons_commons_lang3_jar
 rectangle "spotbugs-annotations\n\n4.8.2" as com_github_spotbugs_spotbugs_annotations_jar
 rectangle "error_prone_annotations\n\n2.23.0" as com_google_errorprone_error_prone_annotations_jar
 rectangle "streamex\n\n0.8.2" as one_util_streamex_jar
-rectangle "codingstyle\n\n3.28.0" as edu_hm_hafner_codingstyle_jar
+rectangle "codingstyle\n\n3.30.0" as edu_hm_hafner_codingstyle_jar
 com_fasterxml_jackson_core_jackson_databind_jar -[#000000]-> com_fasterxml_jackson_core_jackson_annotations_jar
 com_fasterxml_jackson_core_jackson_databind_jar -[#000000]-> com_fasterxml_jackson_core_jackson_core_jar
 edu_hm_hafner_autograding_model_jar -[#000000]-> com_fasterxml_jackson_core_jackson_databind_jar
